backend/requirements.txt
python-dotenv
psycopg2-binary
django
djangorestframework
django-cors-headers
jwt
passlib
django-rest-knox